update origin directory conversion tests
